Se restructura el proyecto para que sea más legible. Falta revisar que todo funcione OK
